Update about.md
We've gotten several customers asking around local routing in MZRs when connected to a DC instead of a PoP. The wording on this page wasn't very clear to them. I've changed the wording a bit and added in a link to make sure the customers understand what is covered by routing in a local market and what is not.

Thank you,

Mathijs Dubbe
Sr. Cloud Design Engineer
mdubbe@nl.ibm.com | mdubbe@softlayer.com

IBM Cloud